+++ docs/CHANGES.md
@@ -24,6 +24,7 @@ should be able to easily backport future new features to older versions rather e
 
 * v12.0.0 (2025-04-16)
   * Added support for Laravel v12.
+  * [BR-256] Fixed `ToArrayConverter` potentially causing fatal error if object lacks `toArray` method.
 
 * v11.0.0 (2024-05-06)
   * **BACKWARD INCOMPATIBLE CHANGES** ([more info](compatibility.md)).